[stateful] Fix state machine for ref-states
It turned out that the state machine worked incorrectly if none of the
transitions had a stateful target state (`function_of_time<T>`). This
patch fixes the problem and adds a unit test for that.

Technical details

Conditional expressions return `ref<T>` if all the branches are
stateless (not functions of time) and `function_of_time<T>` otherwise.
In this case `ref<T>` was implicitly casted to `funciton_of_time<T>`
because it has `operator() (const Time&)`. As the result a snapshot of
the value was taken instead of initialization.