[core] Allow `ref<T>` construction out of values
Adds a new constructor for `ref<T>` class template that initializes the
object as a reference to the newly created constant of the type `T` if
a value convertible to `T` is passed as an argument. This allow calling
functions that expect an expression reference `ref` of concrete type
(where type inference is not required) passing a value without explicit
constant construction (call to `Const()`).