[core] Make `StateMachine` safer
Disables StateMachine (the one that resides in core module) for
aggregate types and, to make sure that the type is correctly marked as
an aggregate, `LiftSelector()` is disabled for regular (non-aggregate)
types.

This change is needed because it is not safe to use field selectors on
'previous' states in the state machine and the solution that has been
implemented before (prohibiting field selectors on previous states)
would not always work.