[internal] Get rid of parallel time counters
Removes duplication of time counters: one in `tick_count`, another one
in `node_time`. Now only the latter remains. As a side effect,
`node_since_activator` now depends on `node_time`, which means that
it will be updated after each 'tick'. This may require optimisation if
the overhead turns out to be significant.